If %html40%, create col elements representing colspecs (implied or explicit), and don't emit widths on each and every cell.
If %html40%, set the tbody valign properly.

If %html40%, allow relative column sizes such as "1*", thus we're not
emitting 50% when we mean 1* (out of 1*, 1*); the effect is that we're
not forcing pgwide if it's not turned on.

Don't set valign on entries, set them on rows (unless they really are
set on the entries).

Some rewriting for speed/efficiency (doens't seem to make much
difference in my timing testing though)