Align refnamediv title correctly when refentry.generate.title is non-zero (closes #1212641).
::Problem:
When refentry.generate.title is non-zero, the title output for
Refnamediv is not aligned flush left.

::Cause:
No code for setting start-indent="" was included in
refentry.title.properties. It should be in order to make the
Refnamediv title output be flush left, as are titles for all other
sectioning children of Refentry

::Fix:
Added code for setting start-indent="" in
refentry.title.properties.