Removed unnecessary trailing comma after final term/glossterm (closes #1215890; thanks to Sam Steingold for reporting the problem).
::PROBLEM::
If a varlistentry or glossentry contains multiple term or
glossterm elements, a comma is rendered after the final term or
glossterm. A comma should instead be rendered only after every
term or glossterm _except_ the last.

::FIX::
Reworked template logic for term/glossterm. They are now handled
with an xsl:for-each in the varlistentry/glossentry template,
rather than as separate templates.

HTML and FO stylesheets appear to have the same problem, so we
probably need to port this change to those as well.