Support man.justify and man.hyphenate params. (closes #1229225).
Note that default for the both of those is zero (off), because
justified text looks good only when it is also hyphenated; to
quote the "Hypenation" node from the groff info page:

  Since the odds are not great for finding a set of words, for
  every output line, which fit nicely on a line without inserting
  excessive amounts of space between words, `gtroff' hyphenates
  words so that it can justify lines without inserting too much
  space between words.

The problem is that groff is not particularly smart about how it
does hyphenation; it can end up hyphenating a lot of things that
you don't want hyphenated (names of symbols, for example), and it
is difficult and tiresome work to prevent it from doing that. So,
disabling both justification and hyphenation ensures that hyphens
won't get inserted where you don't want to them, and you don't end
up with lines containing excessive amounts of space between words.

Yes, these default settings run counter to how most existing man
pages are formatted. But there are some notable exceptions, such
as the perl man pages.