Added the following namespace declarations to the docbookxi.rnc (DocBook 5 + Xinclude) schema:
  namespace src = "http://nwalsh.com/xmlns/litprog/fragment"
  namespace html = "http://www.w3.org/1999/xhtml"

Copied the xsl2/params/refentry.rnc (schema for editing DocBook 5
XSL params litprog files) to the docbook/relaxng/params dir, which
seems like a more appropriate place for it.

Also, changed the refentry.rnc file to include the docbookxi.rnc
(DocBook + Xinclude) schema, so that we can use Xinclude in our
params docs when we want to.