Added new param man.base.url.for.relative.links .. specifies a base URL for relative links (for ulink, @xlink:href, imagedata, audiodata, videodata) shown in the generated NOTES section of man-page output. The value of man.base.url.for.relative.links is prepended to any relative URI that is a value of ulink url, xlink:href, or fileref attribute.
If you use relative URIs in link sources in your DocBook refentry
source, and you leave man.base.url.for.relative.links unset, the
relative links will appear "as is" in the NOTES section of any
man-page output generated from your source. That's probably not
what you want, because such relative links are only usable in the
context of HTML output. So, to make the links meaningful and
usable in the context of man-page output, set a value for
man.base.url.for.relative.links that points
to the online version of HTML output generated from your DocBook
refentry source. For example:

  <xsl:param name="man.base.url.for.relative.links"
  >http://www.kernel.org/pub/software/scm/git/docs/</xsl:param>